dynamic_interaction/ios_dynamic/ios_dynamic_tool/ui_automation/requirements.txt
Appium-Python-Client==2.7.1
pylint==2.5.2
selenium==4.7.2
urllib3==1.26.13
